Changes from last version + map graphs
Hi, 
I made the changes from the last version and add the code for the map graphs for the gene table, I use a world map image instead of libraries like geopandas or geophyton because I though it would be easier and this libraries took a long time to install, and I don't know how common is to have them already installed unlike pandas, numpy and Matplotlib that are the ones used. 
Thanks and sorry for the delay.